+++ .gitignore
@@ -1,3 +1,28 @@
-/* files to be ignored by github*/
-package-lock.json
-node_modules
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/client/node_modules
+/client/.pnp
+/client/.pnp.js
+/client/package-lock.json
+/server/node_modules
+/server/package-lock.json
+
+# testing
+/client/coverage
+
+# production
+
+/client/build
+
+# misc
+
+/client/.DS_Store
+/client/.env.local
+/client/.env.development.local
+/client/.env.test.local
+/client/.env.production.local
+
+/client/npm-debug.log*
+/client/yarn-debug.log*
+/client/yarn-error.log*
