+++ client/src/Common/Footer/Footer.js
@@ -1,0 +1,9 @@
+import React from 'react';
+const Footer = () =>{
+    return(
+        <>
+
+        </>
+    )
+}
+export default Footer;
