+++ client/src/Common/Header/Header.js
@@ -1,0 +1,29 @@
+import React from 'react'
+import { Link } from 'react-router-dom';
+function Header() {
+  return (
+    <nav className="navbar navbar-expand-lg navbar-dark bg-dark">
+	  <div className="container-fluid">
+		<Link className="navbar-brand" to="/">Chit - Chat</Link>
+		<button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+		  <span className="navbar-toggler-icon"></span>
+		</button>
+		<div className="collapse navbar-collapse " id="navbarSupportedContent">
+		  <ul className="navbar-nav ms-auto">
+			<li className="nav-item">
+			  <Link className="nav-link active" aria-current="page" to="/">Home</Link>
+			</li>
+			<li className="nav-item">
+			  <Link className="nav-link" to="/login">Login</Link>
+			</li>
+			<li className="nav-item">
+			  <Link className="nav-link" to="/register">Signup</Link>
+			</li>
+		  </ul>
+		</div>
+	  </div>
+	</nav>
+  )
+}
+
+export default Header
