+++ client/src/Messaging/Header/Header.js
@@ -1,0 +1,8 @@
+import React from "react";
+const Header = ()=>{
+    return(
+        <>
+        </>
+    )
+}
+export default Header;
