+++ client/src/Messaging/RightPane/Messages/Messages.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import '../../Styles/messaging.css'
+const Messages = () =>{
+    return(
+        <div id="messages" className="p-1">
+            
+        </div>
+    )
+}
+export default Messages;
