--- package.json
@@ -1,30 +1,0 @@
-{
-  "name": "chit-chat",
-  "version": "1.0.0",
-  "description": "A chat app for all.",
-  "main": "app.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js"
-  },
-  "author": "Aayush Joshi",
-  "license": "ISC",
-  "dependencies": {
-    "body-parser": "^1.19.1",
-    "cookie-session": "^2.0.0",
-    "cors": "^2.8.5",
-    "crypto": "^1.0.1",
-    "ejs": "^3.1.6",
-    "express": "^4.17.2",
-    "express-session": "^1.17.2",
-    "http": "^0.0.1-security",
-    "method-override": "^3.0.0",
-    "mongoose": "^6.2.1",
-    "multer": "^1.4.4",
-    "passport": "^0.5.2",
-    "passport-local": "^1.0.0",
-    "passport-local-mongoose": "^6.1.0",
-    "socket.io": "^4.4.1",
-    "socket.io-client": "^4.4.1"
-  }
-}
