--- public/stylesheets/messaging.css
@@ -1,128 +1,0 @@
-html{
-    height: 100%;
-}
-body{
-    background-color: #121b20;
-    height: 100%;
-    padding: 2rem;
-}
-@media only screen and (max-width: 800px) {
-    body {
-      padding: inherit;
-    }
-    #main-container,.panels,.sub-container{
-        min-height: 100vh;
-        display: flex;
-        flex-direction: column;
-    }    
-
-    #contacts{
-        flex-grow: 1;
-    }
-    #messages{
-        flex-grow: 1;
-    }    
-}
-a{
-    text-decoration: none;
-}
-.header{
-    background-color: #3949AB;
-}
-#main-container{
-    background-color: #BBF68c;
-    min-height: 90vh;
-    z-index: 1;
-}
-#contacts{
-    overflow-y: scroll;
-    overflow-x: hidden;
-    height: 72.70vh;
-}
-#messages{
-    margin: 0;
-    padding: 0;
-    overflow-y: scroll;
-    overflow-x: hidden;
-    height: 74vh;
-}
-col{
-    padding: 0;
-}
-.card{
-    width: 100%;
-    margin: 0;
-    border-radius: 0;
-}
-#contacts{
-    background-color: #ffffff;
-}
-.color-green{
-    background-color: #81C784;
-}
-.msg{
-    width: auto;
-    max-width: 45%;
-}
-.contact{
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    background-color: blue;
-}
-.contact .name-last-time{
-    display: flex;
-    flex-direction: row;
-}
-.contact .name-last-time {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    background-color: red;
-}
-.name-last-time .name{
-    flex-grow: 1;
-    width: 90%;
-}
-.contact .column{
-    display: flex;
-    flex-direction: column;
-}
-.contact-last-msg{
-        white-space: nowrap;
-        overflow: hidden;
-        display: inline-block;
-        text-overflow: ellipsis;
-        font-size: small;
-        margin: 0;
-}
-.footer-sent{
-    display: flex;
-    flex-direction: row;
-}
- .footer-sent input[text]{
-    flex-grow: 1;
- }
-.footer-sent button{
-    margin-left: 10px;
-}
-
-/* width */
-::-webkit-scrollbar {
-    width: 10px;
-  }
-  
-  /* Track */
-  ::-webkit-scrollbar-track {
-    background: #f1f1f1; 
-  }
-   
-  /* Handle */
-  ::-webkit-scrollbar-thumb {
-    background: #888; 
-  }
-  
-  /* Handle on hover */
-  ::-webkit-scrollbar-thumb:hover {
-    background: #555; 
-  }
