--- views/about/about.ejs
@@ -1,127 +1,0 @@
-<html>
-    <head>
-        <!-- Required meta tags -->
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <!-- Bootstrap CSS -->
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.3.0/socket.io.js"></script>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-        <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" rel="stylesheet">
-        <link rel="stylesheet" type="text/css" href="/stylesheets/about.css">
-        <title>Chit - Chat</title>
-    </head>
-<body>
-     <div class="p-4 header">
-            </div>	
-        <div id="about-main-container" class="container-fluid" style="background-color: white;">
-           
-            <div class="row about-container-row">    
-                <div class="col-md-4 h-100 py-5 d-flex justify-content-center">
-                    <div class=" mx-auto text-center">
-                        <div class="img-thumbnail rounded">
-                            <img class="img-fluid rounded" src="<%=user.image?'.\\'+user.image:'https://winaero.com/blog/wp-content/uploads/2019/11/Photos-new-icon.png'%>">
-                        </div>
-                        <h1 class="display-5 text-center mt-2"><%= user.firstName +" "+user.lastName%></h1>
-                        <p class="display-5 text-center mt-2" style="font-size: larger;">@<%=user.username%></p>
-                        <div class="mt-2">
-                            <a class="btn btn-success btn-lg" href="/logout">Logout</a>
-                            <a href="/messaging" class="btn btn-danger btn-lg">Close</a>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-md-8 h-100 p-4 py-5">
-                    <div class="row">
-                        <ul class="nav nav-tabs">
-                            <li class="nav-item">
-                                <a id="details-container-trigger" class="nav-link active" aria-current="page" href="#">Details</a>
-                            </li>
-                            <li class="nav-item">
-                                <a id="friends-container-trigger" class="nav-link" href="#">Friends</Details></a>
-                            </li>
-                        </ul>
-                    </div>	
-                    <div id="details-container" class="container">
-                        <form class="mt-4" action="" method="post">
-                            <h1 class="display-5 text-center">User Details Page</h1>
-                            <div class="form-group m-2 mt-4">
-                                <input type="text" class="form-control"  required name="firstName" value="<%=user.firstName%>">
-                            </div>
-                            <div class="form- m-2">
-                                <input type="text" class="form-control" required name="lastName" value="<%=user.lastName%>">
-                            </div>
-                            <div class="form-group m-2">
-                                <input type="email" class="form-control" required name="email" value="<%=user.email%>">
-                            </div>
-                            <div class="form-group m-2">
-                                <input type="text" class="form-control" required name="username" value="<%=user.username%>">
-                            </div>
-                            <div class="text-center mt-3">
-                                <button type="submit" class="btn btn-primary btn-lg">Update Data</button>
-                            </div>
-                        </form>
-                    </div>
-                    <div id="friends-container" class="container" >
-                        <div  class="row">
-                            <input id="friend-add-search-text" type="text" class="form-control my-3" placeholder="Search a friend by user Handle to add..." >
-                            <div style="overflow-x:hidden;overflow-y: scroll;height: 50vh;">
-                                <div class="table-responsive" >
-                                    <table id="friend-add-search-table" class="table table-striped" width="100%">
-                                        <col style="width:10%">
-                                        <col style="width:20%">
-                                        <col style="width:20%">
-                                        <col style="width:20%">
-                                        <col style="width:10%">
-                                        <thead>
-                                            <tr>
-                                                <th scope="col">#</th>
-                                                <th scope="col">Name</th>
-                                                <th scope="col">Email</th>
-                                                <th scope="col">Handle</th>
-                                                <th scope="col">Add</th>
-                                            </tr>
-                                        </thead>
-                                        <tbody id="friends-search-body">
-                                            
-                                        </tbody>
-                                    </table>
-                                </div>		
-                                <h3 class="h3">Current Friends</h3>
-                                <hr>
-                                <div class="row">
-                                    <div class="table-responsive">
-                                        <table class="table table-striped" width="100%">
-                                            <col style="width:10%">
-                                            <col style="width:20%">
-                                            <col style="width:20%">
-                                            <col style="width:20%">
-                                            <col style="width:10%">
-                                            <thead>
-                                                <tr>
-                                                    <th scope="col">#</th>
-                                                    <th scope="col">Name</th>
-                                                    <th scope="col">Email</th>
-                                                    <th scope="col">Handle</th>
-                                                    <th scope="col">Remove</th>
-                                                </tr>
-                                            </thead>
-                                            <tbody id="friends-table-body">
-                                                
-                                            </tbody>
-                                        </table>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    
-    <script type="text/javascript">
-		var user_id = "<%= user.id %>";
-		var IP = "<%= ip %>";
-	</script>
-	<script src="/javascript/about.js">  
-	</script>
-</body>
-</html>
