--- views/index.ejs
@@ -1,10 +1,0 @@
-<%-include("partials/header")%>
-    <div class="container-fluid"> 
-      <div class="container" id="landing-header">
-          <h1 class="display-1" id="header">Chit - Chat</h1>
-          <p class="lead" id="header-li">A chatting app for all...</p>
-          <hr>
-          <a href="/register" class="btn btn-light btn-lg">Sign Up</a>
-      </div> 
-    </div>
-<%-include("partials/footer")%>
