--- views/messaging/index.ejs
@@ -1,76 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-	<!-- Required meta tags -->
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1">
-	<meta name="mobile-web-app-capable" content="yes">
-	<meta name="apple-mobile-web-app-capable" content="yes" /> 
-	<!-- Bootstrap CSS -->
-	
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.3.0/socket.io.js"></script>
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-	<link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" rel="stylesheet">
-	<link rel="stylesheet" type="text/css" href="/stylesheets/messaging.css">
-	<title>Chit - Chat</title>
-</head>
-<body>
-	
-	<div  class="container-fluid" id="main-container">
-		<div class="sub-container row">
-			<div class="panels col-md-4 m-0 p-0" >
-				<a href="/about">
-					<div class="card p-2 header">
-						<img src="<%=user.image?'.\\'+user.image:'https://winaero.com/blog/wp-content/uploads/2019/11/Photos-new-icon.png'%>" class="rounded-circle" width="42px" height="42px">
-					</div>
-				</a>
-				<div class="card p-3">
-					<input id="friend-search" type="text" class="form-control p-1 px-3 rounded-pill" placeholder="Enter friend name to search...">
-				</div>
-				<div id="contacts">
-					
-					
-				</div>
-			</div>
-			<div id="right-pane" class="panels col-md-8 p-0 m-0">	
-				<div id="right-pane-div" class="card p-2 header">
-					<div class="row">
-						<div id="back-button" class="col-1 me-0 pe-0 pt-1">
-							<a href="#main-container"><i class="fa-solid fa-arrow-left" style="font-size:xx-large;color: black;"></i></a>
-						</div>
-						<div class="col-2" style="width: 42px;" >
-							<img src="https://winaero.com/blog/wp-content/uploads/2019/11/Photos-new-icon.png" class="rounded-circle" width="42px" height="42px">
-						</div>
-						<div class="col-9 ms-2 py-0" style="color: white;">
-							<div style="font-size: small; font-weight: bold;" id="friend-name">
-								
-							</div>
-							<div style="font-size:small;">
-								Online
-							</div>
-						</div>
-					</div>
-				</div>
-				<div id="messages" class="p-1">
-					
-				</div>
-				<div class="card p-2 m-0">
-					<div class="row">
-						<div class="footer-sent">
-							<input id="sendText" type="text" class="form-control p-2 rounded-pill " placeholder="Enter the text to be sent...">
-							<button id="sendButton" class="btn btn-success float-right">Send</button>
-						</div>
-					</div>
-				</div>
-			</div>
-			
-		</div>
-	</div>
-	<script type="text/javascript">
-		var user_id = "<%= user.id %>";
-		var IP = "<%= ip %>";
-	</script>
-	<script src="/javascript/script.js">  
-	</script>
-</body>
-</html>
