--- views/user/login.ejs
@@ -1,22 +1,0 @@
-<%- include("../partials/header") %>
-	<div class="container">
-		<div class="row">
-			<div class="col-sm-10 col-md-6" style="margin:30px auto;">
-				<div class="p-4" style="background-color:rgba(225,225,225,0.4);">
-					<form action="/login" method="post">
-							<h1 class="display-1 m-2" style="text-align:center;">Login</h1>
-							<div class="form-group m-2">
-								<input type="text" class="form-control" placeholder="username" required name="username">
-							</div>
-							<div class="form-group m-2">
-								<input type="password" class="form-control" placeholder="Password" required name="password">
-							</div>
-							<div class="text-center">
-								<button type="submit" class="btn-lg btn-primary btn-block">Login</button>
-							</div>
-					</form>
-				</div>
-			</div>
-		</div>
-	</div>
-<%-include("../partials/footer")%>
