--- views/user/new.ejs
@@ -1,46 +1,0 @@
-<%- include("../partials/header") %>
-	<div class="container">
-		<div class="row">
-			<div class="col-sm-10 col-md-6" style="margin:30px auto; margin-bottom: 20px;">
-				<div class="p-2" style="background-color:rgba(225,225,225,0.4);">
-					<form action="/register" method="post" enctype="multipart/form-data">
-							<h1 class="display-1 m-2" style="text-align:center;">Sign Up</h1>
-							<div class="form-group m-2 text-center">
-								<img id="blah" src="https://w7.pngwing.com/pngs/831/88/png-transparent-user-profile-computer-icons-user-interface-mystique-miscellaneous-user-interface-design-smile-thumbnail.png"  onclick="document.getElementById('image').click();" alt="your image" style="height: 200px; width: 200px; border-radius: 50%;"/>
-								<input id="image" type="file" style="display: none;" name="userImage"/>
-							</div>		
-							<div class="form-group m-2">
-								<input type="text" class="form-control" placeholder="First Name" required name="firstName">
-							</div>
-							<div class="form-group m-2">
-								<input type="text" class="form-control" placeholder="Last Name" required name="lastName">
-							</div>
-							<div class="form-group m-2">
-								<input type="email" class="form-control" placeholder="Enter email" required name="email">
-							</div>
-							<div class="form-group m-2">
-								<input type="text" class="form-control" placeholder="username" required name="username">
-							</div>
-							<div class="form-group m-2">
-								<input type="password" class="form-control" placeholder="Password" required name="password">
-							</div>
-							<div class="form-group m-2">
-								<input type="password" class="form-control" placeholder="Password" required>
-							</div>
-							<div class="text-center">
-								<button type="submit" class="btn-lg btn-primary btn-block">Signup</button>
-							</div>
-					</form>
-				</div>
-			</div>
-		</div>
-	</div>
-	<script>
-		document.getElementById("image").onchange = evt => {
-			const [file] = document.getElementById("image").files
-			if (file) {
-				blah.src = URL.createObjectURL(file)
-			}
-		}
-	</script>
-<%-include("../partials/footer")%>
